# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AMS - Advanced Medical Stomatology System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List all patients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
  /patients/{id}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get a patient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientDetail"
        "404":
          description: Not found
    patch:
      operationId: updatePatient
      tags: [patients]
      summary: Update a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      operationId: deletePatient
      tags: [patients]
      summary: Delete a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /patients/{id}/ai-recommendations:
    get:
      operationId: getPatientAiRecommendations
      tags: [patients]
      summary: Get AI clinical recommendations for a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiRecommendation"

  # APPOINTMENTS
...
  /treatments:
    get:
      operationId: listTreatments
      tags: [treatments]
      summary: List treatments
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Treatments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Treatment"
    post:
      operationId: createTreatment
      tags: [treatments]
      summary: Create treatment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreatmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
  /treatments/{id}:
    patch:
      operationId: updateTreatment
      tags: [treatments]
      summary: Update treatment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TreatmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Treatment"
    delete:
      operationId: deleteTreatment
      tags: [treatments]
      summary: Delete treatment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PAYMENTS
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, firstName, lastName, phone, dateOfBirth, gender, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female]
        bloodType:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, emergency]
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
        lastVisit:
          type: ["string", "null"]
    PatientDetail:
      allOf:
        - $ref: "#/components/schemas/Patient"
        - type: object
          properties:
            address:
              type: ["string", "null"]
            allergies:
              type: array
              items:
                type: string
            medications:
              type: array
              items:
                type: string
            medicalHistory:
              type: ["string", "null"]
            doctorNotes:
              type: ["string", "null"]
            isSmoker:
              type: boolean
            hasDiabetes:
              type: boolean
            hasHypertension:
              type: boolean
            insuranceProvider:
              type: ["string", "null"]
            insuranceNumber:
              type: ["string", "null"]
            emergencyContact:
              type: ["string", "null"]
            emergencyPhone:
              type: ["string", "null"]
    PatientInput:
      type: object
      required: [firstName, lastName, phone, dateOfBirth, gender]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        dateOfBirth:
          type: string
        gender:
          type: string
          enum: [male, female]
        bloodType:
          type: string
        address:
          type: string
        allergies:
          type: array
          items:
            type: string
        medications:
          type: array
          items:
            type: string
        medicalHistory:
          type: string
        isSmoker:
          type: boolean
        hasDiabetes:
          type: boolean
        hasHypertension:
          type: boolean
        insuranceProvider:
          type: string
        insuranceNumber:
          type: string
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
    PatientUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, inactive, emergency]
        address:
          type: string
        allergies:
          type: array
          items:
            type: string
        medications:
          type: array
          items:
            type: string
        medicalHistory:
          type: string
        doctorNotes:
          type: string
        isSmoker:
          type: boolean
        hasDiabetes:
          type: boolean
        hasHypertension:
          type: boolean
        emergencyContact:
          type: string
        emergencyPhone:
          type: string
    AiRecommendation:
      type: object
      required: [id, type, title, description, priority]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [periodontal, radiographic, medical_alert, treatment, preventive]
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
...
    Treatment:
      type: object
      required: [id, patientId, name, status, startDate, cost]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [planned, in_progress, completed, cancelled]
        startDate:
          type: string
        completedDate:
          type: ["string", "null"]
        cost:
          type: number
        toothNumbers:
          type: array
          items:
            type: integer
        doctorId:
          type: ["integer", "null"]
        doctorName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    TreatmentInput:
      type: object
      required: [patientId, name, cost]
      properties:
        patientId:
          type: integer
        name:
          type: string
        description:
          type: string
        cost:
          type: number
        toothNumbers:
          type: array
          items:
            type: integer
        doctorId:
          type: integer
        notes:
          type: string
    TreatmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planned, in_progress, completed, cancelled]
        completedDate:
          type: string
        cost:
          type: number
        notes:
          type: string
    Payment:
      type: object
      required: [id, patientId, patientName, amount, paidAmount, status, dueDate, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        treatmentId:
          type: ["integer", "null"]
        treatmentName:
          type: ["string", "null"]
        amount:
          type: number
        paidAmount:
          type: number
        status:
          type: string
          enum: [pending, partial, paid, overdue, cancelled]
        dueDate:
          type: string
        createdAt:
          type: string
        notes:
          type: ["string", "null"]
    PaymentInput:
      type: object
      required: [patientId, amount, dueDate]
      properties:
        patientId:
          type: integer
        treatmentId:
          type: integer
        amount:
          type: number
        dueDate:
          type: string
        notes:
          type: string
    PaymentUpdate:
      type: object
      properties:
        paidAmount:
          type: number
        status:
          type: string
          enum: [pending, partial, paid, overdue, cancelled]
        notes:
          type: string
...