# lib/api-zod/src/generated/types/treatmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AMS - Advanced Medical Stomatology System API
 * OpenAPI spec version: 0.1.0
 */

export interface TreatmentInput {
  patientId: number;
  name: string;
  description?: string;
  cost: number;
  toothNumbers?: number[];
  doctorId?: number;
  notes?: string;
}